Refactoring/cleanup/docs/PR feedback pass
