+++ src/db_create_column_family_with_import.md
@@ -1,0 +1,15 @@
+Creates a new column family by importing a collection of SST files. The external SST files can be created in one of the
+following ways:
+
+1. Directly using [`SstFileWriter`](crate::SstFileWriter).
+2. Exported from a particular column family in an existing DB using [`Checkpoint::ExportColumnFamily`][export_cf]. In
+   that case, `metadata` should be the output of [`ExportColumnFamily`][export_cf].
+
+The parameter `import_options` can specify whether to copy or move the external files (default is to copy). If set to
+copy, managing the source files is the caller's responsibility. When set to move, the operation will attempt to delete
+the external files on successful import, logging any failure to delete rather than returning an error. Files are not
+modified on any error, and a best effort is made to remove any newly-created files. A new Column Family will be present
+on a successful return, and will not be present on error. A Column Family may be present if a crash occurs during a
+call.
+
+[export_cf]: crate::checkpoint::Checkpoint::export_column_family
