+++ tests/test_checkpoint.rs
@@ -17,9 +17,10 @@ mod util;
 use pretty_assertions::assert_eq;
 use std::path::Path;
 
-use rust_rocksdb::metadata::ExportImportFilesMetaData;
-use rust_rocksdb::IteratorMode;
-use rust_rocksdb::{checkpoint::Checkpoint, ImportColumnFamilyOptions, Options, DB};
+use rust_rocksdb::checkpoint::Checkpoint;
+use rust_rocksdb::{
+    ExportImportFilesMetaData, ImportColumnFamilyOptions, IteratorMode, Options, DB,
+};
 use util::DBPath;
 
 #[test]
@@ -163,7 +164,7 @@ pub fn test_export_checkpoint_column_family() {
     });
 
     let mut import_metadata = ExportImportFilesMetaData::default();
-    import_metadata.set_db_comparator_name("leveldb.BytewiseComparator");
+    import_metadata.set_db_comparator_name(&export_metadata.get_db_comparator_name());
     import_metadata.set_files(&export_files.to_vec());
 
     let cf_opts = Options::default();
